+++ CHANGELOG.md
@@ -10,6 +10,7 @@ the release.
 - Application Insights exporter improvements - now understands http attributes
   and process links, annotations and messages.
 - ASP.NET Core collector now uses `http.route` for the span name.
+- Initial implementation of Resource Specification.
 
 ## 0.1.0-alpha-33381